core/GameState:
Added a hack to reduce the precision of a y axis collision correction,
which prevents over-correction and causes some strange bugs when the
player tries to walk between a floor and ceiling, with a vertical
gap exactly equal to the player's height.

I should figure out a more permanent and less hackish fix for this.