Switched all of the rendering to OpenGL
The native lib is still hard-coded to be the OSX one, so I'll have
to fix that. I also need to commit the lwjgl jar and native libs
(include their license and readme files).

User input is non-existent, so I'll need to rewrite that part of
the game code to use lwjgl's input handling.